revert(audio): remove Windows capture-event notification changes
