+++ src/server/audio_service/audio_capture_queue.rs
@@ -8,18 +8,13 @@ use std::{
     collections::VecDeque,
     sync::{
         atomic::{AtomicBool, AtomicUsize, Ordering},
-        Arc, Mutex, TryLockError,
+        Arc, Mutex, OnceLock, TryLockError,
     },
-    thread::JoinHandle,
+    thread::{JoinHandle, Thread},
 };
-#[cfg(not(target_os = "windows"))]
-use std::{sync::OnceLock, thread::Thread};
 
 #[path = "audio_capture_encoder.rs"]
 mod encoder;
-#[cfg(target_os = "windows")]
-#[path = "audio_capture_wake.rs"]
-mod wake;
 
 const CAPTURE_PCM_QUEUE_PACKETS: usize = 10;
 const CAPTURE_ENCODER_THREAD_NAME: &str = "audio-encoder";
@@ -64,10 +59,7 @@ impl CapturePcmStats {
 
 struct CapturePcmHandoff {
     buffers: Mutex<CapturePcmBuffers>,
-    #[cfg(not(target_os = "windows"))]
     wake_thread: OnceLock<Thread>,
-    #[cfg(target_os = "windows")]
-    wake: wake::CaptureWake,
     other_dropped: AtomicUsize,
     contention_dropped: AtomicUsize,
     oversized: AtomicUsize,
@@ -99,8 +91,19 @@ pub(super) struct CaptureEncoderConfig {
 pub(super) struct CaptureEncoderWorker {
     stop: Arc<AtomicBool>,
     handle: Option<JoinHandle<()>>,
-    #[cfg(target_os = "windows")]
-    handoff: Arc<CapturePcmHandoff>,
+}
+
+impl Drop for CaptureEncoderWorker {
+    fn drop(&mut self) {
+        self.stop.store(true, Ordering::Release);
+        if let Some(handle) = self.handle.take() {
+            // Rust 1.75's Win7 unpark() can wait for the worker; this owner thread joins it anyway.
+            handle.thread().unpark();
+            if let Err(error) = handle.join() {
+                log::error!("Failed to join audio encoder thread: {error:?}");
+            }
+        }
+    }
 }
 
 struct CaptureEncoderContext {
@@ -122,10 +125,7 @@ pub(super) fn new_pcm_handoff(
             available: Vec::with_capacity(capacity),
             ready: VecDeque::with_capacity(capacity),
         }),
-        #[cfg(not(target_os = "windows"))]
         wake_thread: OnceLock::new(),
-        #[cfg(target_os = "windows")]
-        wake: wake::CaptureWake::new().context("Failed to create audio capture notification")?,
         other_dropped: AtomicUsize::new(0),
         contention_dropped: AtomicUsize::new(0),
         oversized: AtomicUsize::new(0),
@@ -150,7 +150,6 @@ pub(super) fn new_pcm_handoff(
 }
 
 impl CapturePcmSender {
-    #[cfg(not(target_os = "windows"))]
     pub(super) fn set_wake_thread(&self, thread: Thread) -> Result<()> {
         if self.handoff.wake_thread.set(thread).is_err() {
             bail!("Audio capture PCM wake thread is already configured");
@@ -209,9 +208,10 @@ impl CapturePcmSender {
         })
     }
 
-    #[cfg(not(target_os = "windows"))]
     fn wake(&self) {
         if let Some(thread) = self.handoff.wake_thread.get() {
+            // Rust 1.75's Win7 keyed-event wake can wait for the worker to enter its wait.
+            // Capture callbacks therefore do not have a nonblocking guarantee on Win7.
             thread.unpark();
         }
     }
@@ -278,15 +278,11 @@ pub(super) fn start_capture_encoder(
         .name(CAPTURE_ENCODER_THREAD_NAME.to_owned())
         .spawn(move || encoder::run_capture_encoder(context, config))
         .with_context(|| "Failed to start audio encoder thread")?;
-    #[cfg(not(target_os = "windows"))]
     let wake_thread = handle.thread().clone();
     let worker = CaptureEncoderWorker {
         stop,
         handle: Some(handle),
-        #[cfg(target_os = "windows")]
-        handoff: sender.handoff.clone(),
     };
-    #[cfg(not(target_os = "windows"))]
     sender.set_wake_thread(wake_thread)?;
     Ok((sender, worker))
 }
