+++ src/server/audio_service/audio_capture_queue_tests.rs
@@ -59,7 +59,6 @@ fn assert_callback_progress(point: PausePoint) {
         resume: resume_rx,
     };
     let worker = std::thread::spawn(move || paused_worker(receiver, point, pause));
-    #[cfg(not(target_os = "windows"))]
     sender.set_wake_thread(worker.thread().clone()).unwrap();
     entered_rx.recv_timeout(TEST_TIMEOUT).unwrap();
     let (completed_tx, completed_rx) = mpsc::channel();
